Reorganize the react app.
